// lib/notes/interfaces/NoteListSortMode.ts
export enum NoteListSortMode {
  CREATION_DATE_ASCENDING = "CREATION_DATE_ASCENDING",
  CREATION_DATE_DESCENDING = "CREATION_DATE_DESCENDING",
  UPDATE_DATE_ASCENDING = "UPDATE_DATE_ASCENDING",
  UPDATE_DATE_DESCENDING = "UPDATE_DATE_DESCENDING",
  TITLE_ASCENDING = "TITLE_ASCENDING",
  TITLE_DESCENDING = "TITLE_DESCENDING",
  NUMBER_OF_LINKS_ASCENDING = "NUMBER_OF_LINKS_ASCENDING",
  NUMBER_OF_LINKS_DESCENDING = "NUMBER_OF_LINKS_DESCENDING",
  NUMBER_OF_FILES_ASCENDING = "NUMBER_OF_FILES_ASCENDING",
  NUMBER_OF_FILES_DESCENDING = "NUMBER_OF_FILES_DESCENDING",
  NUMBER_OF_CHARACTERS_ASCENDING = "NUMBER_OF_CHARACTERS_ASCENDING",
  NUMBER_OF_CHARACTERS_DESCENDING = "NUMBER_OF_CHARACTERS_DESCENDING",
}