// lib/notes/interfaces/ErrorMessage.ts
export enum ErrorMessage {
  GRAPH_NOT_FOUND = "GRAPH_NOT_FOUND",
  NOTE_NOT_FOUND = "NOTE_NOT_FOUND",
  FILE_NOT_FOUND = "FILE_NOT_FOUND",
  INVALID_MIME_TYPE = "INVALID_MIME_TYPE",
  NOTE_WITH_SAME_TITLE_EXISTS = "NOTE_WITH_SAME_TITLE_EXISTS",
  INVALID_NOTE_STRUCTURE = "INVALID_NOTE_STRUCTURE",
  UNAUTHORIZED = "UNAUTHORIZED",
  INVALID_FILENAME_EXTENSION = "INVALID_FILENAME_EXTENSION",
  NOT_SUPPORTED_BY_STORAGE_PROVIDER = "NOT_SUPPORTED_BY_STORAGE_PROVIDER",
}